requirements.txt
Django==5.1.6
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
